feat: Implement search query capture and forward match (Step 10.1-10.2)
- Added SEARCH mode to editor modes
- Capture search pattern in search_buffer with  key
- Implement buffer_find_pattern() to search forward from start
- Implement buffer_index_to_screen() to convert buffer index to cursor position
- Cursor jumps to first match when Enter is pressed
- Display 'Pattern found' or 'Pattern not found' messages
- Changed cursor_x and cursor_y to size_t for type consistency

Step 10 Search - 2/5 tick boxes complete (capture query, find forward match)

+++ src/editor.h
@@ -64,8 +64,8 @@ typedef enum
 } EditorMode;
 
 typedef struct {
-    int cursor_x;     // Cursor column position
-    int cursor_y;     // Cursor row position
+    size_t cursor_x;     // Cursor column position
+    size_t cursor_y;     // Cursor row position
     size_t screen_rows;  // Terminal height
     size_t screen_cols;  // Terminal width
     size_t row_offset; // Line in the file at the top of the screen
